assignment mongo_db/requirements.txt
cryptography==39.0.0
fastapi~=0.89.1
jwt-signature-validator==0.0.1
orjson==3.8.1
paho-mqtt==1.5.0
pre-commit==2.20.0
pycryptodome==3.16.0
pydantic~=1.10.4
PyJWT==2.4.0
pymongo~=3.13.0
python-dotenv~=0.21.0
redis==3.5.3
uvicorn==0.17.5
